Update README.md
This is my first C program written during BSc Computer Science. The purpose of this program is to print the numbers from 1 to 10 using a for loop.

The program begins with including the standard input/output header file <stdio.h>. Inside the main() function, a for loop is used. The loop starts with i = 1, continues as long as i <= 10, and increments i by 1 in each iteration. During every iteration, the value of i is printed on the screen using printf().

Thus, the output of the program will be the sequence of numbers from 1 to 10, each printed on a new line.